python/requirements.txt
google-cloud-core==2.4.1
google-api-core==2.21.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.26.0
google-cloud-storage==2.18.2
pycryptodome==3.21.0
boto3==1.35.45
psycopg2-binary==2.9.9
pandas==2.2.1
pyarrow==17.0.0